+++ test/SanitizersTest.php
@@ -1,14 +1,17 @@
 <?php
+use Sanitizers\Sanitizers\Sanitizer;
+
+const EOL = PHP_EOL . PHP_EOL;
+
 $currentDir = dirname(__FILE__);
 $baseDir = dirname($currentDir);
 if (is_readable($baseDir . "/vendor/autoload.php")) {
     require_once $baseDir . "/vendor/autoload.php";
+    echo "Using composer" . EOL;
 } else {
     require_once $baseDir . "/src/Sanitizers.php";
+    echo "Not using composer" . EOL;
 }
-use Sanitizers\Sanitizers\Sanitizer;
-
-const EOL = PHP_EOL . PHP_EOL;
 
 $sanitize = new Sanitizer(false);
 $len = 32; //32 bytes = 256 bits
@@ -24,6 +27,8 @@ if (function_exists("random_bytes"))
 }
 $test_values = [
     "hex" => bin2hex($bytes),
+    "int" => $len-0.5,
+    "float" => $len-0.5,
     "name" => "\0Some Name ä\x80",
     "email" => "ADMIN@ExAMpLe.com",
     "message" => "Hi <script>alert('I am XSS');</script>",
@@ -31,6 +36,8 @@ $test_values = [
 ];
 $values = [
     "hex" => $sanitize->Hex($test_values["hex"]),
+    "int" => $sanitize->Integer($test_values["int"]),
+    "float" => $sanitize->Float($test_values["int"]),
     "name" => $sanitize->Name($test_values["name"]),
     "email" => $sanitize->Email($test_values["email"]),
     "message" => $sanitize->clean($test_values["message"], false, true),