+++ test/SanitizersTest.php
@@ -75,7 +75,8 @@ $testValues = array(
     "username" => "PuneetGopinath", // It will become to smaller case if you want upper case also then use sanitize function with type parameter as name e.g. `$sanitizer->sanitize("name", $username)`
     "html" => "<b>Text in bold</b><!-- This is a comment --><link rel=stylesheet src=http://ha.ckers.org/bad.css /><a href=\"javascript:alert('XSS');\">XSS</a>",
     "password" => "\$UnIQUe|`_-<script>alert('XSS')</script>#pass•WorD%!?",
-    "function_clean" => "XSS <script>alert('XSS');</script>"
+    "function_clean" => "XSS <script>alert('XSS');</script>",
+    "function_escape" => "Sanitizers\"; DELETE FROM users; ---"
 );
 
 if ($configFromIni && is_readable($baseDir . "/src/config.ini"))
@@ -94,7 +95,8 @@ $values = array(
     "username" => $sanitizer->sanitize("username", $testValues["username"]),
     "html" => $sanitizer->HTML($testValues["html"]),
     "password" => $sanitizer->sanitize("password", $testValues["password"]),
-    "function_clean" => $sanitizer->clean($testValues["function_clean"])
+    "function_clean" => $sanitizer->clean($testValues["function_clean"]),
+    "function_escape" => $sanitizer->escape($testValues["function_escape"])
 );
 $filters = array(
     "types" => array(
@@ -107,7 +109,8 @@ $filters = array(
         "url" => "url",
         "username" => "username",
         "html" => "html",
-        "function_clean" => "" //Will use clean function
+        "function_clean" => "", //Will use clean function
+        "function_escape" => "escape"
     ),
     "message" => array(
         "trim" => false, //Enables php trim function, default:true
@@ -124,13 +127,13 @@ $auto_values = $sanitizer->sanitizeArray($testValues, $filters);
 echo "Array Key -- Original Value => Sanitized Value -- same: bool" . EOL;
 
 foreach ($testValues as $i => $value) {
-    echo $i . " -- " . $value . " => " . $values[$i] . " -- same: " . (($value === $values[$i])?"true":"false") . EOL;
+    echo $i . " -- \"" . $value . "\" => \"" . $values[$i] . "\" -- same: " . (($value === $values[$i])?"true":"false") . EOL;
 }
 
 echo EOL . "Array Key -- Original Value => Auto Sanitized Value -- same: bool" . EOL;
 
 foreach ($testValues as $i => $value) {
-    echo $i . " -- " . $value . " => " . $auto_values[$i] . " -- same: " . (($value === $auto_values[$i])?"true":"false") . EOL;
+    echo $i . " -- \"" . $value . "\" => \"" . $auto_values[$i] . "\" -- same: " . (($value === $auto_values[$i])?"true":"false") . EOL;
 }
 
 if ($debug)